+++ src/Attributes/DataProvider.php
@@ -7,7 +7,7 @@ use Attribute;
 
 /**
  * Data provider attribute
- * Defines a data source for parameters of a test
+ * Defines a data source for parameters of a test, it is name of a method in the same class
  *
  * @author Jakub Konečný
  */