+++ src/DataProvider.php
@@ -15,6 +15,8 @@ final readonly class DataProvider implements IDataProvider
 {
     public const string ANNOTATION_NAME = "dataProvider";
 
+    public const string ANNOTATION_EXTERNAL_NAME = "dataProviderExternal";
+
     public function __construct(private Reader $annotationsReader)
     {
     }
@@ -29,6 +31,7 @@ final readonly class DataProvider implements IDataProvider
         if ($reflection->getNumberOfParameters() < 1) {
             return [];
         }
+
         $dataProvider = $this->annotationsReader->getAnnotation(self::ANNOTATION_NAME, $class, $method);
         if (is_string($dataProvider)) {
             $className = $reflection->getDeclaringClass()->getName();
@@ -51,6 +54,36 @@ final readonly class DataProvider implements IDataProvider
                 throw new InvalidDataProviderException("Method $className::$dataProvider does not exist.", 0, $e);
             }
         }
+
+        $dataProviderExternal = $this->annotationsReader->getAnnotation(
+            self::ANNOTATION_EXTERNAL_NAME,
+            $class,
+            $method
+        );
+        if (is_string($dataProviderExternal)) {
+            try {
+                [$className, $methodName] = explode("::", $dataProviderExternal);
+                $reflection = new ReflectionMethod($className, $methodName);
+                if (!$reflection->isPublic()) {
+                    throw new InvalidDataProviderException("Method $dataProviderExternal is not public.");
+                }
+                if (!$reflection->isStatic()) {
+                    throw new InvalidDataProviderException("Method $dataProviderExternal is not static.");
+                }
+                /** @var string&callable $dataProviderExternal */
+                $result = call_user_func($dataProviderExternal);
+                if (!is_iterable($result)) {
+                    throw new InvalidDataProviderException(
+                        "Method $dataProviderExternal has to return an array or an iterable object."
+                    );
+                }
+                /** @var iterable[] $result */
+                return $result;
+            } catch (\ReflectionException $e) {
+                throw new InvalidDataProviderException("Method $dataProviderExternal does not exist.", 0, $e);
+            }
+        }
+
         return [];
     }
 }