+++ tests/Annotations/PhpAttributesEngineTest.php
@@ -3,8 +3,11 @@ declare(strict_types=1);
 
 namespace MyTester\Annotations;
 
+use MyTester\Attributes\DataProviderExternal;
 use MyTester\Attributes\Skip;
 use MyTester\Attributes\TestSuite;
+use MyTester\DataProvider;
+use MyTester\ExternalDataProvider;
 use MyTester\SkipChecker;
 use MyTester\TestCase;
 
@@ -32,10 +35,19 @@ final class PhpAttributesEngineTest extends TestCase
         $this->assertSame("PhpAttributesEngine", $engine->getAnnotation(TestCase::ANNOTATION_TEST_SUITE, self::class));
         $this->assertNull($engine->getAnnotation(TestCase::ANNOTATION_TEST_SUITE, self::class, "method"));
         $this->assertSame("", $engine->getAnnotation(SkipChecker::ANNOTATION_NAME, self::class, "method"));
+        $this->assertSame(
+            ExternalDataProvider::class . "::dataProviderArray",
+            $engine->getAnnotation(DataProvider::ANNOTATION_EXTERNAL_NAME, self::class, "dataProviderExternal")
+        );
     }
 
     #[Skip()]
     private function method(): void
+    {
+    }
+
+    #[DataProviderExternal(ExternalDataProvider::class, "dataProviderArray")]
+    private function dataProviderExternal(): void
     {
     }
 }