+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 8.0.0-dev
 - data providers can now also return iterable objects not just arrays
 - added option to suppress warning about no performed assertions in a test method/test case
 - possible BC break: attribute Test, TestSuite and DataProvider are now readonly
+- added support for external data providers
 
 Version 7.3.1
 - allowed installation konecnyjakub/event-dispatcher 2